server/pyproject.toml
[tool.poetry]
name = "OrdoAbChaos"
version = "0.1.0"
description = ""
authors = ["Yartsev Vladimir <v.rainovsky@yandex.ru>"]
readme = "README.md"
packages = [
    { include = "src" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.0"
uvicorn = {extras = ["standart"], version = "^0.21.1"}
sqlalchemy = "^2.0.7"
asyncmy = "^0.2.7"
isoweek = "^1.3.3"
loguru = "^0.7.0"
strawberry-graphql = {extras = ["debug"], version = "^0.171.1"}
aiosqlite = "^0.18.0"

[tool.poetry.scripts]
#recreate_tables = "scripts:recreate_tables"
run_server = "server:run"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
